Remove tagging in grammar. See msg.
Turns out Pest has a bug (see
https://github.com/pest-parser/pest/issues/875) where it fails to detect
left-recursion when tags are used.

If we remove all the tags, then it is able to successfully detect left
recursion better.

This commit removes the tags, and also fixes up one instance of
left-recursion.

We still have the badly left-recursive `bin_expr` to fix up, but that is
left for another commit.